requirements/dev-requirements.txt
-r requirements.txt
-r test-requirements.txt

bandit
black
django-sslserver
flake8
ipython
pre-commit
pytest
pytest-cov
pytest-django
pytest-mock
pytest-rerunfailures
safety
-e git+https://github.com/jsumnerPhD/http-observatory#egg=httpobs
